+++ packages/api/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.1](https://github.com/sapphire-project/plugins/compare/@sapphire/plugin-api@1.3.0...@sapphire/plugin-api@1.3.1) (2021-01-23)
+
+### Bug Fixes
+
+-   **plugin-api:** return null instead of throwing error ([#46](https://github.com/sapphire-project/plugins/issues/46)) ([7a0cf3d](https://github.com/sapphire-project/plugins/commit/7a0cf3d6d135ae4ab5de72e9e0c22446b3a9fffd))
+
+# [1.3.0](https://github.com/sapphire-project/plugins/compare/@sapphire/plugin-api@1.2.0...@sapphire/plugin-api@1.3.0) (2021-01-09)
+
+### Bug Fixes
+
+-   **api:** update to latest discord-api-types ([6a46797](https://github.com/sapphire-project/plugins/commit/6a4679770f44b4930016280963c0fe4563c48bf2))
+
+### Features
+
+-   **api:** add media parsers ([#39](https://github.com/sapphire-project/plugins/issues/39)) ([18fac10](https://github.com/sapphire-project/plugins/commit/18fac107e9210a8c66da6b7cc5d3b24ab47a07e1))
+
 # [1.2.0](https://github.com/sapphire-project/plugins/compare/@sapphire/plugin-api@1.1.1...@sapphire/plugin-api@1.2.0) (2020-12-28)
 
 ### Bug Fixes